type fix

+++ inc/libft.h
@@ -101,7 +101,7 @@ void		ft_striteri(char *s, void (*f)(unsigned int, char*));
 void		ft_formatSpaces(char *s);
 void		str_sed(char **base, char *ind, char *add);
 char		*ft_strmerge(char *s1, char *s2);
-char		*ft_strsmerge(int n, ...);
+char		*ft_strsmerge(size_t n, ...);
 
 /*
 	Printing
